WIP (remove SignUpForm component)
