+++ src/app/auth/components/SignUp/SignUp.tsx
@@ -1,26 +1,35 @@
-import { useState, useMemo } from 'react';
-import { SubmitHandler } from 'react-hook-form';
-import { Helmet } from 'react-helmet-async';
-import localStorageService, { STORAGE_KEYS } from 'app/core/services/local-storage.service';
+import { useState, useEffect, useMemo } from 'react';
+import { SubmitHandler, useForm, useWatch } from 'react-hook-form';
 import queryString from 'query-string';
+import { auth } from '@internxt/lib';
+import { Link } from 'react-router-dom';
+import { Info, WarningCircle } from '@phosphor-icons/react';
+import { Helmet } from 'react-helmet-async';
+import localStorageService, { STORAGE_KEYS } from '../../../core/services/local-storage.service';
 
-import { useAppDispatch } from 'app/store/hooks';
-import { userActions, userThunks } from 'app/store/slices/user';
-import { planThunks } from 'app/store/slices/plan';
-import errorService from 'app/core/services/error.service';
-import navigationService from 'app/core/services/navigation.service';
-import { productsThunks } from 'app/store/slices/products';
-import { AppView, IFormValues } from 'app/core/types';
-import { referralsThunks } from 'app/store/slices/referrals';
+import { useAppDispatch } from '../../../store/hooks';
+import { userActions, userThunks } from '../../../store/slices/user';
+import { planThunks } from '../../../store/slices/plan';
+import errorService from '../../../core/services/error.service';
+import navigationService from '../../../core/services/navigation.service';
+import { productsThunks } from '../../../store/slices/products';
+import { AppView, IFormValues } from '../../../core/types';
+import { referralsThunks } from '../../../store/slices/referrals';
+import TextInput from '../TextInput/TextInput';
+import PasswordInput from '../PasswordInput/PasswordInput';
+import testPasswordStrength from '@internxt/lib/dist/src/auth/testPasswordStrength';
+import PasswordStrengthIndicator from '../../../shared/components/PasswordStrengthIndicator';
 import { useSignUp } from './useSignUp';
 import { UserSettings } from '@internxt/sdk/dist/shared/types/userSettings';
-import authService, { getNewToken } from 'app/auth/services/auth.service';
+import { useTranslationContext } from '../../../i18n/provider/TranslationProvider';
+import authService, { getNewToken } from '../../../auth/services/auth.service';
 import PreparingWorkspaceAnimation from '../PreparingWorkspaceAnimation/PreparingWorkspaceAnimation';
-import paymentService from 'app/payment/services/payment.service';
-import { decryptPrivateKey } from 'app/crypto/services/keys.service';
-import analyticsService from 'app/analytics/services/analytics.service';
-import SignupForm from './SignupForm';
-import DownloadBackupKey from 'app/auth/views/DownloadBackupKey/DownloadBackupKey';
+import paymentService from '../../../payment/services/payment.service';
+import { MAX_PASSWORD_LENGTH } from '../../../shared/components/ValidPassword';
+import { decryptPrivateKey } from '../../../crypto/services/keys.service';
+import analyticsService from '../../../analytics/services/analytics.service';
+import Button from '../../../shared/components/Button/Button';
+import DownloadBackupKey from '../../../auth/views/DownloadBackupKey/DownloadBackupKey';
 
 export interface SignUpProps {
   location: {
@@ -36,6 +45,10 @@ interface ViewsProps {
 }
 
 function SignUp(props: SignUpProps): JSX.Element {
+  const { translate } = useTranslationContext();
+  const [isValidPassword, setIsValidPassword] = useState(false);
+  const [view, setView] = useState<Views>('signUp');
+
   const qs = queryString.parse(navigationService.history.location.search);
   const autoSubmit = useMemo(
     () => authService.extractOneUseCredentialsForAutoSubmit(new URLSearchParams(window.location.search)),
@@ -46,14 +59,86 @@ function SignUp(props: SignUpProps): JSX.Element {
     qs.register === 'activate' ? 'activate' : 'appsumo',
     hasReferrer ? String(qs.ref) : undefined,
   );
+  const hasEmailParam = (qs.email && auth.isValidEmail(qs.email as string)) || false;
 
+  const getInitialEmailValue = () => {
+    if (hasEmailParam) {
+      return qs.email as string;
+    }
+
+    if (autoSubmit.enabled && autoSubmit.credentials?.email) {
+      return autoSubmit.credentials.email;
+    }
+  };
+  const {
+    register,
+    formState: { errors, isValid },
+    handleSubmit,
+    control,
+    getValues,
+  } = useForm<IFormValues>({
+    mode: 'onChange',
+    defaultValues: {
+      email: getInitialEmailValue(),
+      password: autoSubmit.enabled && autoSubmit.credentials ? autoSubmit.credentials.password : '',
+    },
+  });
   const dispatch = useAppDispatch();
+  const password = useWatch({ control, name: 'password', defaultValue: '' });
   const [signupError, setSignupError] = useState<Error | string>();
   const [showError, setShowError] = useState(false);
   const [isLoading, setIsLoading] = useState(false);
-  const [view, setView] = useState<Views>('signUp');
-
+  const [passwordState, setPasswordState] = useState<{
+    tag: 'error' | 'warning' | 'success';
+    label: string;
+  } | null>(null);
+  const [showPasswordIndicator, setShowPasswordIndicator] = useState(false);
   const showPreparingWorkspaceAnimation = useMemo(() => autoSubmit.enabled && !showError, [autoSubmit, showError]);
+
+  const formInputError = Object.values(errors)[0];
+
+  let bottomInfoError: null | string = null;
+
+  if (formInputError?.message) {
+    bottomInfoError = formInputError.message;
+  } else if (showError && signupError) {
+    bottomInfoError = signupError.toString();
+  }
+
+  useEffect(() => {
+    if (autoSubmit.enabled && autoSubmit.credentials) {
+      onSubmit(getValues());
+    }
+  }, []);
+
+  useEffect(() => {
+    if (password.length > 0) onChangeHandler(password);
+  }, [password]);
+
+  function onChangeHandler(input: string) {
+    setIsValidPassword(false);
+    if (input.length > MAX_PASSWORD_LENGTH) {
+      setPasswordState({ tag: 'error', label: translate('modals.changePasswordModal.errors.longPassword') });
+      return;
+    }
+
+    const result = testPasswordStrength(input, (qs.email as string) === undefined ? '' : (qs.email as string));
+
+    setIsValidPassword(result.valid);
+    if (!result.valid) {
+      setPasswordState({
+        tag: 'error',
+        label:
+          result.reason === 'NOT_COMPLEX_ENOUGH'
+            ? 'Password is not complex enough'
+            : 'Password has to be at least 8 characters long',
+      });
+    } else if (result.strength === 'medium') {
+      setPasswordState({ tag: 'warning', label: 'Password is weak' });
+    } else {
+      setPasswordState({ tag: 'success', label: 'Password is strong' });
+    }
+  }
 
   const onSubmit: SubmitHandler<IFormValues> = async (formData, event) => {
     const redeemCodeObject = autoSubmit.credentials && autoSubmit.credentials.redeemCodeObject;
@@ -128,16 +213,90 @@ function SignUp(props: SignUpProps): JSX.Element {
     }
   };
 
+  const getLoginLink = () => {
+    const currentParams = new URLSearchParams(window.location.search);
+
+    return currentParams.toString() ? '/login?' + currentParams.toString() : '/login';
+  };
+
   const View = ({ view }: ViewsProps) => {
     const views = {
       signUp: (
-        <SignupForm
-          onSubmit={onSubmit}
-          isLoading={isLoading}
-          signupError={signupError}
-          autoSubmit={autoSubmit}
-          showError={showError}
-        />
+        <div className="flex flex-col items-start space-y-5">
+          <h1 className="text-3xl font-medium">{translate('auth.signup.title')}</h1>
+
+          <form className="flex w-full flex-col space-y-2" onSubmit={handleSubmit(onSubmit)}>
+            <TextInput
+              placeholder={translate('auth.email') as string}
+              label="email"
+              type="email"
+              disabled={hasEmailParam}
+              register={register}
+              required={true}
+              minLength={{ value: 1, message: 'Email must not be empty' }}
+              error={errors.email}
+            />
+
+            <label className="space-y-0.5">
+              <PasswordInput
+                className={passwordState ? passwordState.tag : ''}
+                placeholder={translate('auth.password')}
+                label="password"
+                maxLength={MAX_PASSWORD_LENGTH}
+                register={register}
+                onFocus={() => setShowPasswordIndicator(true)}
+                required={true}
+                error={errors.password}
+              />
+              {showPasswordIndicator && passwordState && (
+                <PasswordStrengthIndicator className="pt-1" strength={passwordState.tag} label={passwordState.label} />
+              )}
+              {bottomInfoError && (
+                <div className="flex flex-row items-start pt-1">
+                  <div className="flex h-5 flex-row items-center">
+                    <WarningCircle weight="fill" className="mr-1 h-4 text-red" />
+                  </div>
+                  <span className="font-base w-56 text-sm text-red">{bottomInfoError}</span>
+                </div>
+              )}
+            </label>
+
+            <div className="flex space-x-2.5 rounded-lg bg-primary/10 p-3 pr-4 dark:bg-primary/20">
+              <Info size={20} className="shrink-0 text-primary" />
+              <p className="text-xs">
+                Internxt doesn't store passwords.{' '}
+                <span className="font-semibold">
+                  In case you forget your password, you will lose access to all your files.
+                </span>
+              </p>
+            </div>
+
+            <Button disabled={isLoading || !isValidPassword} loading={isLoading} variant="primary" className="w-full">
+              {isValid && isValidPassword
+                ? `${translate('auth.signup.encrypting')}...`
+                : translate('auth.signup.title')}
+            </Button>
+          </form>
+
+          <span className="mt-2 w-full text-xs text-gray-50">
+            {translate('auth.terms1')}{' '}
+            <a href="https://internxt.com/legal" target="_blank" className="text-xs text-gray-50 hover:text-gray-60">
+              {translate('auth.terms2')}
+            </a>
+          </span>
+
+          <div className="w-full border-b border-gray-10" />
+
+          <div className="flex w-full items-center justify-center space-x-1.5 font-medium">
+            <span>{translate('auth.signup.haveAccount')}</span>
+            <Link
+              to={getLoginLink()}
+              className="cursor-pointer font-medium text-primary no-underline hover:text-primary focus:text-primary-dark"
+            >
+              {translate('auth.signup.login')}
+            </Link>
+          </div>
+        </div>
       ),
       downloadBackupKey: <DownloadBackupKey />,
     };
